refactor(webhooks): split handler helpers [OPE-408]
